fix: removed types file
