--- example/storybook/src/example-screens/dashboard/types.ts
@@ -1,12 +1,0 @@
-interface CommentCardProps {
-  userName: string;
-  comment: string;
-}
-
-interface SidebarItemProps {
-  tooltipText: string;
-  icon: React.JSX.Element;
-  itemProps: any;
-}
-
-export { CommentCardProps, SidebarItemProps };
